4단계 ë서비스 모두 구현

+++ src/main/java/lotto/model/Rank.java
@@ -30,7 +30,7 @@ public enum Rank {
         return winningMoney;
     }
 
-    public static Rank valueOf(long countOfMatch, boolean matchBonus) {
+    public static Rank valueOf(int countOfMatch, boolean matchBonus) {
         if (countOfMatch < 3) {
             return Rank.MISS;
         }
